[docs] Use Next.js primitives for head metadata
instead of manually rendering html.
