+++ docs/src/app/(public)/(content)/layout.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import type { Metadata, Viewport } from 'next/types';
+import type { Metadata } from 'next/types';
 import * as SideNav from 'docs/src/components/SideNav';
 import * as QuickNav from 'docs/src/components/QuickNav/QuickNav';
 import { Header } from 'docs/src/components/Header';
@@ -38,28 +38,4 @@ export default function Layout({ children }: React.PropsWithChildren) {
 export const metadata: Metadata = {
   title: null,
   description: null,
-};
-
-export const viewport: Viewport = {
-  themeColor: [
-    // Desktop Safari header background
-    {
-      media: '(prefers-color-scheme: light) and (min-width: 1024px)',
-      color: 'oklch(95% 0.25% 264)',
-    },
-    {
-      media: '(prefers-color-scheme: dark) and (min-width: 1024px)',
-      color: 'oklch(25% 1% 264)',
-    },
-
-    // Mobile Safari header background (match the site header)
-    {
-      media: '(prefers-color-scheme: light)',
-      color: 'oklch(98% 0.25% 264)',
-    },
-    {
-      media: '(prefers-color-scheme: dark)',
-      color: 'oklch(17% 1% 264)',
-    },
-  ],
 };
