+++ docs/src/app/layout.tsx
@@ -1,18 +1,39 @@
 import * as React from 'react';
-import { Metadata } from 'next';
-import { Favicons } from './Favicons';
+import type { Metadata, Viewport } from 'next';
 
 export default function Layout({ children }: React.PropsWithChildren) {
   return (
     <html lang="en">
-      <head>
-        <meta name="viewport" content="initial-scale=1, width=device-width" />
-        <Favicons />
-      </head>
       <body>{children}</body>
     </html>
   );
 }
+
+export const viewport: Viewport = {
+  initialScale: 1,
+  width: 'device-width',
+  themeColor: [
+    // Desktop Safari page background
+    {
+      media: '(prefers-color-scheme: light) and (min-width: 1024px)',
+      color: 'oklch(95% 0.25% 264)',
+    },
+    {
+      media: '(prefers-color-scheme: dark) and (min-width: 1024px)',
+      color: 'oklch(25% 1% 264)',
+    },
+
+    // Mobile Safari header background (match the page)
+    {
+      media: '(prefers-color-scheme: light)',
+      color: '#FFF',
+    },
+    {
+      media: '(prefers-color-scheme: dark)',
+      color: '#000',
+    },
+  ],
+};
 
 export const metadata: Metadata = {
   title: {
@@ -31,5 +52,31 @@ export const metadata: Metadata = {
       default: 'Base UI',
     },
     ttl: 604800,
+  },
+  icons: {
+    icon: [
+      {
+        rel: 'icon',
+        url:
+          process.env.NODE_ENV !== 'production'
+            ? '/static/favicon-dev.ico'
+            : '/static/favicon.ico',
+        sizes: '32x32',
+      },
+      {
+        rel: 'icon',
+        type: 'svg+xml',
+        url:
+          process.env.NODE_ENV !== 'production'
+            ? '/static/favicon-dev.svg'
+            : '/static/favicon.svg',
+      },
+    ],
+    apple: [
+      {
+        rel: 'apple-touch-icon',
+        url: '/static/apple-touch-icon.png',
+      },
+    ],
   },
 };
